# Optimize array_constructor